% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{check_displaymatrix}
\alias{check_displaymatrix}
\title{Internal helper function: Check if there attribute coding matches}
\usage{
check_displaymatrix(md_define = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}
}
\value{
NULL. Throws a warning if not matched
}
\description{
Internal helper function: Check if there attribute coding matches
}
