% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{check_response_number}
\alias{check_response_number}
\title{Internal helper function: Check incomplete response per respondent}
\usage{
check_response_number(md_define = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}
}
\value{
A list containing the data and processed data
}
\description{
Internal helper function: Check incomplete response per respondent
}
