% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate.R
\name{plot_aggregate_betas}
\alias{plot_aggregate_betas}
\title{Aggregate plot and save data}
\usage{
plot_aggregate_betas(md_define = NULL, label_width = 30, point_color = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{An integer. Capping label length}

\item{point_color}{A string denoting the color of the plotted data}
}
\value{
A list containing the data and processed data
}
\description{
Aggregate plot and save data
}
\examples{
plot_aggregate_betas(md_define, label_width = 30)
}
