% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate.R
\name{plot_aggregate_prct}
\alias{plot_aggregate_prct}
\title{Aggregate plot and save data}
\usage{
plot_aggregate_prct(md_define = NULL, label_width = 30, point_color = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{An integer. Capping label length}

\item{point_color}{A string denoting the color of the plotted data}
}
\value{
A list containing the data and processed data
}
\description{
Aggregate plot and save data
}
\examples{
file_path <- system.file("extdata", "icecream.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
md_define <- read_qualtrics_header(md_define = md_define)
md_define <- process_attribute_text(md_define = md_define)
md_define <- process_maxdiff_question_format(md_define = md_define, signchange = FALSE)
md_define <- process_bayesm_format(md_define = md_define)
md_define <- run_bayesm_simulation(md_define, R = 5e2)
md_define <- process_aggregate_prct(md_define)
md_define <- plot_aggregate_prct(md_define, label_width = 30)
}
