% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{plot attribute correlations}
\usage{
plot_corr(md_define = NULL, n_clusters = 1)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{n_clusters}{An integer. Capping label length}
}
\value{
A list containing the data and processed data
}
\description{
plot attribute correlations
}
\examples{
plot_corr(md_define, n_clusters = 3)
}
