% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{plot_count}
\alias{plot_count}
\title{Count plot}
\usage{
plot_count(md_define = NULL, label_width = 30)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{maximum width of text labels}
}
\value{
A list containing the data and processed data
}
\description{
Count plot
}
\examples{
plot_aggregate_betas(md_define)
}
