% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group.R
\name{plot_group_prct}
\alias{plot_group_prct}
\title{Group plot and save data}
\usage{
plot_group_prct(md_define, label_width = 30, include = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{maximum width of text labels}

\item{include}{The groups to include. Must match exact wording in data.}
}
\value{
A list containing the data and processed data
}
\description{
Group plot and save data
}
\examples{
plot_group_prct(md_define, label_width = 30, include = c("Female", "Male", "Unspecified"))
}
