% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlogit.R
\name{plot_mlogit_betas}
\alias{plot_mlogit_betas}
\title{Plotting \code{mlogit} results for checking}
\usage{
plot_mlogit_betas(md_define, label_width = 30, point_color = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{An integer. Capping label length}

\item{point_color}{A string denoting the color of the plotted data}
}
\value{
A list containing the data and processed data
}
\description{
Plotting \code{mlogit} results for checking
}
\examples{
file_path <- system.file("extdata", "icecream.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
md_define <- read_qualtrics_header(md_define = md_define)
md_define <- process_attribute_text(md_define = md_define)
md_define <- process_maxdiff_question_format(md_define = md_define, signchange = FALSE)
md_define <- run_mlogit(md_define)
md_define <- plot_mlogit_betas(md_define, label_width = 30, point_color = c("red", "blue"))
}
