% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesm_simulation.R
\name{plot_trace}
\alias{plot_trace}
\title{Making trace plot for the MCMC draws}
\usage{
plot_trace(md_define = NULL, label_width = 30)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{An integer. Capping label length}
}
\value{
A list containing the data and processed data
}
\description{
Making trace plot for the MCMC draws
}
\examples{
plot_trace(md_define)
}
