% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate.R
\name{process_aggregate_betas}
\alias{process_aggregate_betas}
\title{process aggregate betas}
\usage{
process_aggregate_betas(md_define, burn_in = 0.5)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{burn_in}{burn in period before the simulated data are used}
}
\value{
A list containing the data and processed data
}
\description{
process aggregate betas
}
\examples{
md_define <- read_qualtrics_data("inst/extdata/sample_data.csv")
md_define <- read_qualtrics_header(md_define = md_define)
md_define <- process_attribute_text(md_define = md_define)
md_define <- process_maxdiff_question_format(md_define = md_define, signchange = FALSE)
md_define <- process_bayesm_format(md_define = md_define)
md_define <- run_bayesm_simulation(md_define, R = 2e3)
md_define <- process_aggregate_betas(md_define = md_define)
}
