% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate.R
\name{process_aggregate_prct}
\alias{process_aggregate_prct}
\title{Post simulation analysis: Percentage}
\usage{
process_aggregate_prct(md_define = NULL, burn_in = 0.5)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{burn_in}{burn in period before the simulated data are used}
}
\value{
A list containing the data and processed data
}
\description{
Post simulation analysis: Percentage
}
\examples{
process_aggregate_prct(md_define)
}
