% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{process_attribute_text}
\alias{process_attribute_text}
\title{Internal helper function: Format headers of data file}
\usage{
process_attribute_text(md_define = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}
}
\value{
A list containing the data and processed data
}
\description{
Internal helper function: Format headers of data file
}
