% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bayesm_format.R
\name{process_bayesm_format}
\alias{process_bayesm_format}
\title{Convert data to bayesm format}
\usage{
process_bayesm_format(md_define = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}
}
\value{
A list containing the data and processed data
}
\description{
Convert data to bayesm format
}
\examples{
file_path <- system.file("extdata", "icecream.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
md_define <- read_qualtrics_header(md_define = md_define)
md_define <- process_attribute_text(md_define = md_define)
md_define <- process_maxdiff_question_format(md_define = md_define, signchange = FALSE)
md_define <- process_bayesm_format(md_define = md_define)
}
