% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group.R
\name{process_group_prct}
\alias{process_group_prct}
\title{Group data processing}
\usage{
process_group_prct(md_define = NULL, group_var = NULL, burn_in = 0.5)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{group_var}{A string denoting the variable in the data set header}

\item{burn_in}{burn in period before the simulated data are used}
}
\value{
A list containing the data and processed data
}
\description{
Group data processing
}
\examples{
process_group_prct(md_define, group_var = "Q4")
}
