% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual.R
\name{process_individual_prct}
\alias{process_individual_prct}
\title{Individual estimation}
\usage{
process_individual_prct(md_define = NULL, burn_in = 0.5)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{burn_in}{burn in period before the simulated data are used}
}
\value{
A list containing the data and processed data
}
\description{
Individual estimation
}
\examples{
plot_individual_prct(md_define)
}
