% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{process_maxdiff_module_format}
\alias{process_maxdiff_module_format}
\title{Parse maxdiff module format data (instead of maxdiff question format)}
\usage{
process_maxdiff_module_format(md_define = NULL, signchange = FALSE)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{signchange}{Boolean, whether the scale should flip. Default: FALSE}
}
\value{
A list containing the data and processed data
}
\description{
Parse maxdiff module format data (instead of maxdiff question format)
}
\examples{
file_path <- system.file("extdata", "sixfeatures.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
md_define <- process_maxdiff_module_format(md_define, signchange = FALSE)
}
