% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_qualtrics.R
\name{read_qualtrics_data}
\alias{read_qualtrics_data}
\title{Read in a csv file}
\usage{
read_qualtrics_data(file_name = NULL, folder = NULL, md_define = NULL)
}
\arguments{
\item{file_name}{A string denoting a csv file, e.g., "data.csv"}

\item{folder}{A string denoting the path of where the data file is.}

\item{md_define}{A list containing a sublist of data}
}
\value{
A list containing the data and processed data
}
\description{
Read in a csv file
}
\examples{
file_path <- system.file("extdata", "icecream.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
}
