% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_qualtrics.R
\name{read_qualtrics_header}
\alias{read_qualtrics_header}
\title{Process Qualtrics dataset headers}
\usage{
read_qualtrics_header(md_define = NULL, n_word_match = 10)
}
\arguments{
\item{md_define}{A list containing the data and processed data}

\item{n_word_match}{An integer. Number of words match for the maxdiff question}
}
\value{
A list containing the data and processed data
}
\description{
Process Qualtrics dataset headers
}
\examples{
file_path <- system.file("extdata", "icecream.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
md_define <- read_qualtrics_header(md_define = md_define)
}
