% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual.R
\name{plot_attribute_distribution}
\alias{plot_attribute_distribution}
\title{Plot individual distribution}
\usage{
plot_attribute_distribution(md_define, label_width = 30, ncol = 2)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{An integer. Capping label length}

\item{ncol}{Number of column in the plots}
}
\value{
A list containing the data and processed data
}
\description{
Plot individual distribution
}
\examples{
plot_attribute_distribution(md_define)
}
