% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{plot_count}
\alias{plot_count}
\title{Count plot}
\usage{
plot_count(md_define = NULL, label_width = 30)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{maximum width of text labels}
}
\value{
A list containing the data and processed data
}
\description{
Count plot
}
\examples{
file_path <- system.file("extdata", "icecream.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
md_define <- read_qualtrics_header(md_define = md_define)
md_define <- process_attribute_text(md_define = md_define)
md_define <- process_maxdiff_question_format(md_define = md_define, signchange = FALSE)
md_define <- process_bayesm_format(md_define = md_define)
md_define <- plot_count(md_define)
}
