% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlogit.R
\name{plot_mlogit_betas}
\alias{plot_mlogit_betas}
\title{Plotting \code{mlogit} results for checking}
\usage{
plot_mlogit_betas(md_define, label_width = 30, point_color = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{An integer. Capping label length}

\item{point_color}{A string denoting the color of the plotted data}
}
\value{
A list containing the data and processed data
}
\description{
Plotting \code{mlogit} results for checking
}
\examples{
plot_mlogit_betas(md_define, label_width = 30, point_color = c("red", "blue"))
}
