% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{process_attribute_text}
\alias{process_attribute_text}
\title{Format headers of data file}
\usage{
process_attribute_text(md_define = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}
}
\value{
A list containing the data and processed data
}
\description{
Format headers of data file
}
\examples{
file_path <- system.file("extdata", "icecream.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
md_define <- read_qualtrics_header(md_define = md_define)
md_define <- process_attribute_text(md_define = md_define)
}
