% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bayesm_format.R
\name{process_bayesm_format}
\alias{process_bayesm_format}
\title{Convert data to bayesm format}
\usage{
process_bayesm_format(md_define = NULL)
}
\arguments{
\item{md_define}{A list containing a sublist of data}
}
\value{
A list containing the data and processed data
}
\description{
Convert data to bayesm format
}
\examples{
process_bayesm_format(md_define)
}
