% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{process_corr}
\alias{process_corr}
\title{process data to plot attribute correlations}
\usage{
process_corr(md_define = NULL, label_width = 30, burn_in = 0.5)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{label_width}{An integer. Capping label length}

\item{burn_in}{burn in period before the simulated data are used}
}
\value{
A list containing the data and processed data
}
\description{
process data to plot attribute correlations
}
\examples{
plot_corr(md_define)
}
