% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_qualtrics.R
\name{read_qualtrics_data}
\alias{read_qualtrics_data}
\title{Add together two numbers}
\usage{
read_qualtrics_data(file_name = NULL, folder = NULL, md_define = NULL)
}
\arguments{
\item{file_name}{A string denoting a csv file, e.g., "abc.csv"}

\item{folder}{A string denoting the path of where the data file is.}
}
\value{
md_define A list containing the data and processed data
}
\description{
Add together two numbers
}
\examples{
read_qualtrics_data(file_name = "data.csv", folder = "/project/", md_define = md_define)
}
