% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesm_simulation.R
\name{run_bayesm_simulation}
\alias{run_bayesm_simulation}
\title{\code{bayesm} wrapper: Run \code{rhierMnlRwMixture} simulation with the parsed format}
\usage{
run_bayesm_simulation(
  md_define = md_define,
  R = 10000,
  keep = 10,
  seed = NA_integer_
)
}
\arguments{
\item{md_define}{A list containing a sublist of data}

\item{R}{An integer. Number of MCMC draws}

\item{keep}{An integer. Keeping every \code{keep}-th draw. Default: keeping 2000 draws.}

\item{seed}{An integer. Provide a seed for results reproduction}
}
\value{
A list containing the data and processed data
}
\description{
\code{bayesm} wrapper: Run \code{rhierMnlRwMixture} simulation with the parsed format
}
\examples{
file_path <- system.file("extdata", "icecream.csv", package = "gmaxdiff")
md_define <- read_qualtrics_data(file_path)
md_define <- read_qualtrics_header(md_define = md_define)
md_define <- process_attribute_text(md_define = md_define)
md_define <- process_maxdiff_question_format(md_define = md_define, signchange = FALSE)
md_define <- process_bayesm_format(md_define = md_define)
md_define <- run_bayesm_simulation(md_define, R = 5e2)
}
