% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mlogit.R
\name{run_mlogit}
\alias{run_mlogit}
\title{Wrapper for \code{mlogit} for data checking}
\usage{
run_mlogit(md_define)
}
\arguments{
\item{md_define}{A list containing a sublist of data}
}
\value{
A list containing the data and processed data
}
\description{
Wrapper for \code{mlogit} for data checking
}
\examples{
run_mlogit(md_define)
}
